requirements.txt
discord.py
python-dotenv
asyncio
json
time
random
traceback
